back to casting counters

+++ src/hdb.hpp
@@ -340,7 +340,7 @@ int DBG::count_nulls(){
 template<typename T>
 int DBG::delete_if_below_val(T min){
   khint_t k;
-  double class_size = static_cast<double>(size());
+  uint64_t class_size = size();
   uint64_t counter = 0;
   for (k = kh_begin(class_h); k != kh_end(class_h); ++k){  // traverse
     if (kh_exist(class_h, k)){            // test if a bucket contains data
@@ -350,7 +350,7 @@ int DBG::delete_if_below_val(T min){
       counter++;
       if (class_print == 1){
         if ( counter % 20000 == 0){
-          std::cout << "\r" << "   - " << (counter/class_size)*100 << "% (" << counter << " of " << class_size << " )  " << std::flush;
+          std::cout << "\r" << "   - " << ((double)counter/(double)class_size)*100 << "% (" << counter << " of " << class_size << " )  " << std::flush;
         }
       }
     }
@@ -365,7 +365,7 @@ int DBG::delete_if_below_val(T min){
 template<typename T>
 int DBG::delete_if_above_val(T max){
   khint_t k;
-  double class_size = static_cast<double>(size());
+  uint64_t class_size = size();
   uint64_t counter = 0;
   for (k = kh_begin(class_h); k != kh_end(class_h); ++k){  // traverse
     if (kh_exist(class_h, k)){            // test if a bucket contains data
@@ -375,7 +375,7 @@ int DBG::delete_if_above_val(T max){
       counter++;
       if (class_print == 1){
         if ( counter % 20000 == 0){
-          std::cout << "\r" << "   - " << (counter/class_size)*100 << "% (" << counter << " of " << class_size << " )  " << std::flush;
+          std::cout << "\r" << "   - " << ((double)counter/(double)class_size)*100 << "% (" << counter << " of " << class_size << " )  " << std::flush;
         }
       }
     }
@@ -389,7 +389,7 @@ int DBG::delete_if_above_val(T max){
 /* Delete all of the kmers that are flagged for deletion   */
 int DBG::delete_flagged(){
   khint_t k;
-  double class_size = static_cast<double>(size());
+  uint64_t class_size = size();
   uint64_t counter = 0;
   for (k = kh_begin(class_h); k != kh_end(class_h); ++k){  // traverse
     if (kh_exist(class_h, k)){            // test if a bucket contains data
@@ -399,7 +399,7 @@ int DBG::delete_flagged(){
       counter++;
       if (class_print == 1){
         if ( counter % 20000 == 0){
-          std::cout << "\r" << "   - " << (counter/class_size)*100 << "% (" << counter << " of " << class_size << " )  " << std::flush;
+          std::cout << "\r" << "   - " << ((double)counter/(double)class_size)*100 << "% (" << counter << " of " << class_size << " )  " << std::flush;
         }
       }
     }
@@ -417,7 +417,7 @@ int DBG::mark_branching(){
   khint_t k;
   uint64_t key;
   DBnode * pNode;
-  double class_size = static_cast<double>(size());
+  uint64_t class_size = size();
   uint64_t counter = 0;
   uint64_t op_counter = 0;
   uint16_t fiveprime_counter = 0;
@@ -448,7 +448,7 @@ int DBG::mark_branching(){
       op_counter++;
       if (class_print == 1){
         if ( op_counter % 20000 == 0){
-          std::cout << "\r" << "   - " << (op_counter/class_size)*100 << "% (" << op_counter << " of " << class_size << " )  " << std::flush;
+          std::cout << "\r" << "   - " << ((double)op_counter/(double)class_size)*100 << "% (" << op_counter << " of " << class_size << " )  " << std::flush;
         }
       }
     }
@@ -464,7 +464,7 @@ int DBG::mark_branching(){
 int DBG::mark_all_as_unvisited(){
   khint_t k;
   uint64_t key;
-  double class_size = static_cast<double>(size());
+  uint64_t class_size = size();
   uint64_t counter = 0;
   DBnode * pNode;
   for (k = kh_begin(class_h); k != kh_end(class_h); ++k){  // traverse
@@ -475,7 +475,7 @@ int DBG::mark_all_as_unvisited(){
       counter++;
       if (class_print == 1){
         if ( counter % 20000 == 0){
-          std::cout << "\r" << "   - " << (counter/class_size)*100 << "% (" << counter << " of " << class_size << " )  " << std::flush;
+          std::cout << "\r" << "   - " << ((double)counter/(double)class_size)*100 << "% (" << counter << " of " << class_size << " )  " << std::flush;
         }
       }
     }
